Rename recent search and viewed components. Remove underscore.
