+++ pathways_vue/components/search/recent-views.vue
